# backend/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Clover_TH API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/admin/members": {
      "get": {
        "tags": [
          "members"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "0",
                "1",
                "true",
                "false"
              ]
            },
            "in": "query",
            "name": "incomplete",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "0",
                "1",
                "true",
                "false"
              ]
            },
            "in": "query",
            "name": "includeInactive",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "ign": {
                        "type": "string"
                      },
                      "nickname": {
                        "nullable": true,
                        "type": "string"
                      },
                      "jobId": {
                        "type": "number"
                      },
                      "discordId": {
                        "type": "string"
                      },
                      "isActive": {
                        "type": "boolean"
                      },
                      "isAdmin": {
                        "type": "boolean"
                      },
                      "isIncomplete": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "id",
                      "ign",
                      "nickname",
                      "jobId",
                      "discordId",
                      "isActive",
                      "isAdmin",
                      "isIncomplete"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/members/{id}": {
      "patch": {
        "tags": [
          "members"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ign": {
                    "type": "string",
                    "maxLength": 256
                  },
                  "nickname": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 256
                  },
                  "jobId": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true,
                    "maximum": 9007199254740991
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "ign": {
                      "type": "string"
                    },
                    "nickname": {
                      "nullable": true,
                      "type": "string"
                    },
                    "jobId": {
                      "type": "number"
                    },
                    "discordId": {
                      "type": "string"
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "isAdmin": {
                      "type": "boolean"
                    },
                    "isIncomplete": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "ign",
                    "nickname",
                    "jobId",
                    "discordId",
                    "isActive",
                    "isAdmin",
                    "isIncomplete"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/members/{id}/deactivate": {
      "post": {
        "tags": [
          "members"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "ign": {
                      "type": "string"
                    },
                    "nickname": {
                      "nullable": true,
                      "type": "string"
                    },
                    "jobId": {
                      "type": "number"
                    },
                    "discordId": {
                      "type": "string"
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "isAdmin": {
                      "type": "boolean"
                    },
                    "isIncomplete": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "ign",
                    "nickname",
                    "jobId",
                    "discordId",
                    "isActive",
                    "isAdmin",
                    "isIncomplete"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/members/{id}/reactivate": {
      "post": {
        "tags": [
          "members"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "ign": {
                      "type": "string"
                    },
                    "nickname": {
                      "nullable": true,
                      "type": "string"
                    },
                    "jobId": {
                      "type": "number"
                    },
                    "discordId": {
                      "type": "string"
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "isAdmin": {
                      "type": "boolean"
                    },
                    "isIncomplete": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "ign",
                    "nickname",
                    "jobId",
                    "discordId",
                    "isActive",
                    "isAdmin",
                    "isIncomplete"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}